Move CachedTagResult and TagLookupJob to its own component
